+++ CHANGELOG.md
@@ -28,7 +28,7 @@
 - Optimize the trappings search
 - Manage trapping to be equiped/wearing by default
 - \*add deutsch compatibility
-- \*Limit generation button fore those who can create actor
+- Limit generation button fore those who can create actor
 
 ### Bugfix